[One .NET] Fix targeting API 33, build API 31 ref pack
Fixes: #6914

The Microsoft.Android `@(KnownFrameworkReference)` will now declare
targeting and runtime pack metadata that is based on the
`$(TargetPlatformVersion)` of the project file.  The only exception to
this rule is when '$(TargetPlatformVersion)' == '31.0'; in this case the
project will use an API 31 ref pack with API 32 runtime packs.  The
inline `Microsoft.Android.Sdk.BundledVersions.targets` file content
has been moved into a separate file to improve the readability of the
`@(KnownFrameworkReference)` element.

Breakdown of supported `$(TargetFramework)` -> ref/runtime pack used:

| `$(TargetFramework)` | Targeting Pack | Runtime Pack |
| -------------------- | -------------- | ------------ |
| net6.0-android     | Microsoft.Android.Ref.31 | Microsoft.Android.Runtime.32 |
| net6.0-android31.0 | Microsoft.Android.Ref.31 | Microsoft.Android.Runtime.32 |
| net6.0-android32.0 | Microsoft.Android.Ref.32 | Microsoft.Android.Runtime.32 |
| net6.0-android33.0 | Microsoft.Android.Ref.33 | Microsoft.Android.Runtime.33 |

Tests have been updated accordingly, and they will now use a dynamic
test case source that should make sure we don't miss any coverage when
making future updates to `$(AndroidDefaultTargetDotnetApiLevel)`,
`$(AndroidLatestStableApiLevel)`, and `$(AndroidLatestUnstableApiLevel)`.

If an unsupported `$(TargetFramework)` is used in a project file the
following errors will now be displayed:

    "test.csproj" (_GenerateRestoreGraphProjectEntry target) (1:5) ->
       (ProcessFrameworkReferences target) ->
         C:\Program Files\dotnet\sdk\6.0.300-preview.22211.2\Sdks\Microsoft.NET.Sdk\targets\Microsoft.NET.Sdk.FrameworkReferenceResolution.targets(90,5): error : NETSDKZZZZ: Error getting pack version: Pack 'Microsoft.Android.Ref.25' was not present in workload manifests.
         C:\Program Files\dotnet\sdk\6.0.300-preview.22211.2\Sdks\Microsoft.NET.Sdk\targets\Microsoft.NET.Sdk.FrameworkReferenceResolution.targets(90,5): error : NETSDKZZZZ: Error getting pack version: Pack 'Microsoft.Android.Runtime.25.android-arm' was not present in workload manifests.
         C:\Program Files\dotnet\sdk\6.0.300-preview.22211.2\Sdks\Microsoft.NET.Sdk\targets\Microsoft.NET.Sdk.FrameworkReferenceResolution.targets(90,5): error : NETSDKZZZZ: Error getting pack version: Pack 'Microsoft.Android.Runtime.25.android-arm64' was not present in workload manifests.
         C:\Program Files\dotnet\sdk\6.0.300-preview.22211.2\Sdks\Microsoft.NET.Sdk\targets\Microsoft.NET.Sdk.FrameworkReferenceResolution.targets(90,5): error : NETSDKZZZZ: Error getting pack version: Pack 'Microsoft.Android.Runtime.25.android-x86' was not present in workload manifests.
         C:\Program Files\dotnet\sdk\6.0.300-preview.22211.2\Sdks\Microsoft.NET.Sdk\targets\Microsoft.NET.Sdk.FrameworkReferenceResolution.targets(90,5): error : NETSDKZZZZ: Error getting pack version: Pack 'Microsoft.Android.Runtime.25.android-x64' was not present in workload manifests.

Additionally, the build has been updated to produce a reference pack for
the API level declared in `$(AndroidDefaultTargetDotnetApiLevel)`, if it
is not the same as `$(AndroidLatestStableApiLevel)`.  This will allow
the default `net6.0-android` target framework to use an up to date
Mono.Android.dll, instead of an older preview version from NuGet.org.