[Xamarin.Android.Build.Tests] Ignore NoSymbolsArgShouldReduceAppSize (#6907)
Temporarily ignore this test to reduce noise on PR builds while it is
being fixed.  The test has seemingly been partially invalidated by
commit 5efda9d69, however it is still passing against Classic
Xamarin.Android on Windows.  More investigation is required as to why
we see different stripping results when using a combination of the `-s`,
`no-write-symbols`, and `nodebug` flags with different toolchains.