Better timing infrastructure
We rely on precise (nanosecond level) performance measurements and, as
we improve various components of the code, the precision requirements
increase - especially with regards to the amount of indeterminism.  One
such source of indeterminism are the `logcat` calls which we make as the
measurements are taken.  Such calls send the message to a system daemon
which needs to acquire locks in order to put messages in the `logcat`
buffer(s).  Each such call means that the operating system schedules our
process out and an undetermined amount of time can pass before control
comes back to our thread.

In order to avoid this particular issue, this commit implements a new
timing infrastructure which can log events without incurring the cost of
`logcat` until such time as the timing is done and the information can
be logged without affecting measurements.  This is done by introducing a
new timing mode, `timing=fast-bare`, which can be used in the
`debug.mono.log` system property instead of `timing` or `timing=bare`.
The mode is similar to `timing=bare` but it postpones until a broadcast
intent is sent to the application.  This can be done using the following
command:

    adb shell broadcast -a mono.android.app.DUMP_TIMING_DATA <PACKAGE_NAME>

At this point, all the gathered events are logged.  In addition, the new
infrastructure measures the time it spent initializing itself, which can
then be subtracted from the total time to get as close to "real world"
performance of the app as possible.

In testing, I've seen that this causes timing results to be more stable
as well as saving us up to 4% time for the `Displayed` measure.