[Xamarin.Android.Build.Tasks] fix --runtime and r8 at the same time
If you do:

    dotnet new android
    dotnet build -c Release -r android-arm64 -p:AndroidLinkTool=r8

You'll get the warning:

    warning XA4304: ProGuard configuration file 'obj\Release\android-arm64\android-arm64\proguard\proguard_project_references.cfg' was not found.

That path has two `android-arm64` in it!

Additionally, I found this problem was hidden in our testing
infrastructure because we set `$(AppendRuntimeIdentifierToOutputPath)`
to `false` in several tests... After removing this, I could write a
test that reproduces this issue.

The fix for the issue is that `$(_InnerIntermediateOutputPath)` needs
to check if the outer build has a single `$(RuntimeIdentifier)`. If
so, we don't need to append a RID in that path.