Fix SuccessfulAndroidXApp Unit Test (#6958)
The commit bfc6817 was merged with a broken unit test. The problem was with the Nuget restore of the AndroidX packages.

```
error NU1202: Package Xamarin.AndroidX.AppCompat 1.4.1.1 is not compatible with monoandroid50 (MonoAndroid,Version=v5.0).
```

This was very confusing since the TargetFrameworkVersion should have been monoandroid12.0 which is picked up
from Configuration.props. However this was not happening. Further investigation revieled that Configuration.props does not even exist when this test was being run. This is because we build the test assemblies as part of a different job, archive them, and then finally extract and run them. The Configuration.props file is not part of that archive.

So we need to put in a fallback setting for the TargetFrameworkVersion to handle the case when the Configuration.props file does not exist.