[generator] Fix reserved keywords binary search (#977)
Context: a3182bc04f909161d1d4147d13f9746ce4df25e7
Context: https://docs.microsoft.com/en-us/dotnet/core/extensions/performing-culture-insensitive-string-operations-in-arrays

Commit a3182bc0 enabled string code analyzers, because in e.g. the
`sr` locale, `"Ljava".StartsWith("L")` would be *false*, because `Lj`
was a language-specific digraph, and "special-cased".  For intended
behavior, we need to use `.StartsWith("L", StringComparison.Ordinal)`.

There is a ["similar yet different" problem][0] with `Array.Sort()`
and `Array.BinarySearch()` and string arrays: a locale-specific
comparison is performed by default:

	var reserved_keywords = new [] {
	    "as", "catch", "char", "checked", "class", "const",
	};
	var index = Array.BinarySearch (reserved_keywords, "checked");

The *expectation* is that `index` is 3.  In the `cs-CZ` locale, "ch"
is a digraph, and `index` is -7, i.e. was not found.

There is currently no Code Analysis check for this scenario.

Review all `Array.Sort()` and `Array.BinarySearch()` invocations so
that an appropriate `IComparer` instance is provided.  This allows
e.g. `checked` to be properly `@escaped` when created a binding in a
machine with a `cs-CZ` locale.

[0]: https://docs.microsoft.com/en-us/dotnet/core/extensions/performing-culture-insensitive-string-operations-in-arrays