Parse cookies on redirect (#6987)
Fixes: https://github.com/xamarin/xamarin-android/issues/3954

When using the `AllowAutoRedirect` feature of `AndroidMessageHandler` to automatically follow 3xx redirects, 
the cookies that are returned together with the redirection are not processed. If the redirection target requires 
the cookies returned by the first request, the call will fail.

This change fixes this behavior and always processes the cookies before the redirection takes place.