[One .NET] support building net6.0-android apps in .NET 7
We currently have a hard dependency between:

* `Xamarin.Android.Build.Tasks.dll`
* `libmonodroid.so`

The design for building a `net6.0-android` app with a .NET 7 SDK
results in us using a .NET 7 `Xamarin.Android.Builds.Tasks.dll` and a
.NET 6 `libmonodroid.so`. This crashes with:

    05-06 09:32:31.206  8713  8713 E AndroidRuntime: java.lang.UnsatisfiedLinkError: dlopen failed: cannot locate symbol "java_name_width" referenced by "/data/app/~~8CHtXY8wK4g7s9VcOLTQEg==/com.microsoft.net6.helloandroid-cUUhwuQls7TRb6JdBsbPdg==/split_config.arm64_v8a.apk!/lib/arm64-v8a/libmonodroid.so"...

To solve this problem:

* Define `$(AndroidNet6Version)` to specify our .NET 6 GA `32.0.301`
  build that is released.

* Add an alias for a `Microsoft.Android.Sdk.$(HostOS).NET6` workload
  pack that loads a .NET 6 version of `Microsoft.Android.Sdk.$(HostOS)`.

* Include both the .NET 6 and .NET 7 versions of these packs in the
  workload.

* Check `$(TargetFrameworkVersion)` and import the .NET 7 or .NET 6
  MSBuild targets when appropriate.

* Update the `@(KnownFrameworkReference)` to the version of .NET 6
  that our .NET 7 packs know about.

* We have a matching .NET 6 `Xamarin.Android.Build.Tasks.dll` and
  `libmonodroid.so`!

This obviously increases our install footprint of
`Microsoft.Android.Sdk.$(HostOS)` by 2x. We might be able to address
this by splitting out some files into a
`Microsoft.Android.Sdk.Tooling` pack that can be shared between .NET 6
and .NET 7. Perhaps files like `bundletool.jar` and `r8.jar` could be
shared?

One hack I had to put in place was to avoid the crash:

    05-06 12:21:19.981 18868 18868 E AndroidRuntime: java.lang.UnsatisfiedLinkError: dlopen failed: cannot locate symbol "mono_opt_aot_lazy_assembly_load" referenced by "/data/app/~~Fp4gEr_9sxU1qU6PoI-v6Q==/com.companyname.foo-EKEtL67XJI-d0H17_3oz3g==/lib/arm64/libmonodroid.so"...

The .NET 7 SDK we are on, does not yet know about .NET runtime 6.0.5
(it uses 6.0.3), so I manually fixed this version number for now. We
should be able to remove this eventually when the .NET 7 SDK provides
6.0.5 by default.

A second hack I put in place:

* Make Microsoft.Android.Sdk a "framework" and not an "sdk" in
  `WorkloadManifest.json`.

This prevents two `AutoImport.props` from being imported... for the
.NET 7 one and the .NET 6 one.

This seems very weird, but this still works:

    <Import Project="Sdk.targets" Sdk="Microsoft.Android.Sdk" />

We can remove this if I condition everything in `AutoImport.props`
behind a current `$(TargetFrameworkVersion)`. We would then need to
ship this in a .NET 6 release.