Add `_AndroidFastTiming` MSbuild property. This property will control if the `BroadcastReceiver` will be included in the `AndroidManifest.xml` of the application. Without setting this property `fast-timing` data will NOT be collected.
So in addition to calling

`adb shell setprop debug.mono.log timing=fast-bare`

The user must also pass `-p:_AndroidFastTiming=true` to the build OR
set `<_AndroidFastTiming>true</_AndroidFastTiming>` in their csproj.

`dotnet build -t:Install -p:Configuration=Release -p:_AndroidFastTiming=true`

This will register the built in `BroadcastReceiver` and allow it to
respond to the broadcast.

It will default to `` so the user has to explicitly set it.