Address @jonathanpeppers comments.
To *begin* answering the "how do we *test* this?" question,
add a new `@(_AndroidRemapMembers)` build action.

`@(_AndroidMamMappingFile)` is JSON input which is converted into XML
by the `<MamJsonToXml/>` task.  This is the XML described in prior
commits, and in the current PR body text.

`@(_AndroidRemapMembers)` is XML input, of the same format.

This allows us to use our "internal XML" format in "external"
apps -- HelloWorld, future unit tests -- in order to demonstrate
that the remapping infrastructure *works*.

This means we can update HelloWorld (for now, unit tests later)
to have the XML file:

    <replacements>
      <replace-type from="android/app/Activity" to="android/app/ListActivity" />
    </replacements>

and update `HelloWorld.DotNet.csproj` to have:

    <_AndroidRemapMembers Include="remap.xml" />

At runtime, `adb logcat` contains:

    # jonp: found replacement type: `android/app/Activity` => `android/app/ListActivity`

*And*, the app *works*!  No crashes!  (Yay!)