[Xamarin.Android.Build.Tasks] Add %(AndroidJavaSource.Bind) (#5926)
Update the `@(AndroidJavaSource)` Build action:

 1. In .NET 7, all `**\*.java` files are automatically added to the
    `@(AndroidJavaSource)` Build action.

 2. Add `%(AndroidJavaSource.Bind)` item metadata, a boolean value
    (default is True) which controls whether or not the file is
    compiled and bound *prior* to the C# build.

The combination of these means that custom Java code can be added to
any .NET Android project, and a binding of that Java code will be
usable by the C# code.  Both Java code and bindings will be included
in App builds of referencing apps.

There are some restrictions on the Java code.  You can only use Java
types or types which are defined in `.jar` or `.aar` files referenced
by the project via `@(AndroidLibrary)` or `@(EmbeddedJar)`.  It is
suggested that you *do not* have `publi` Java types which use Java
generics.  As with normal bindings, Java generics can cause
significant problems; try to stuck to normal types if you can.