[build] enable the dotnet7 feed (#6998)
We should now have access to use the `dnceng-dotnet7` connection in
DevDiv's AzDO instance.

Now that we'll be pushing each package to a feed in
xamarin-android/main and release branches, we should make a small
correction to the version number.

`$(AndroidPackVersion)` should 33.0.0 be, as we don't have a need to
mirror .NET version bands in our version number. In stable releases we
plan to use the patch version as the "commit distance" like we have
done in .NET 6 GA builds.