[build] Refactor net6 references in src/monodroid (#7001)
Instances of `NET6` in src/monodroid and xaprepare have been removed or
replaced with `NET`.  This should hopefully improve readability and
maintainability as we begin work on .NET 7 and beyond.

Build and output directories for One .NET versions of monodroid.csproj
have been updated to use `$(AndroidRID)` as a unique identifier, rather
than `$(AndroidAbi)-net6`.