[profiled-aot] added a simple way to record AOT profiles
Based on:

* https://github.com/dotnet/maui/tree/main/src/ProfiledAot
* https://github.com/jonathanpeppers/Mono.Profiler.Android#usage-of-the-aot-profiler

To record a new profile, attach an arm64 device and run:

    $ ./bin/Release/dotnet/dotnet build src/profiled-aot/build.proj

`dotnet.aotprofile.txt` is a list of method names contained within the
profile. We don't ship these, but we can use them to track changes
over time. Note that they are not always in order, so I opened the
files in VS Code and did Ctrl+Shift+P to get the command palette. Then
`Sort lines ascending` to get them in alphabetical order. If the text
files don't change, it is likely not necessary to update the
`.aotprofile` files.

Along with these changes, I moved `dotnet.aotprofile` to the
`src/profiled-aot` directory.

I also updated `AotProfileTests` to use:

https://www.nuget.org/packages/Mono.AotProfiler.Android/7.0.0-preview1

Which should work in .NET 7.