Fix remaining unit tests?
Move `TypeAndMemberRemapping()` *earlier* in the file.
`DotNetInstallAndRun()` appears in the **Tests** tab, but
`TypeAndMemberRemapping()` was not, which was very confusing.
On a lark, suspect that `TypeAndMemberRemapping()` isn't run
because it's after `DotNetDebug()`, which is `[Category("Debugger")]`.
Will moving the test fix things?

Fix Mono.Android.NET-Tests, via removing a debug message from
AndroidRuntime, and test fixes within external/Java.Interop.