Fix Windows Build Error on Prepare
We are seeing the Prepare step fail on windows machines on CI. This
is because `dotnet` is interpreting the Errors we are emitting when
downloading the `dotnet` sdk as actual errors. This results in the
`dotnet` process exiting with an ExitCode of -1.

The thing is these are not really errors they are warnings. So lets
just use `LogWarning` instead to get around this issue.