Don't update `@(_JavaSourceJavadocXml)`
Context: https://github.com/xamarin/xamarin-android/pull/7046#issuecomment-1145268477
Context: https://github.com/xamarin/xamarin-android/pull/7046#issuecomment-1146280691

*Why* does the `DotNetBuild()` unit test now fail?  It didn't before.
(It's failing because it's producing warnings, and we assert that no
warnings are produced.)

Warnings are produced, apparently, because there are duplicate items
in `@(_JavaSourceJavadocXml)`, and *one* of the `<JavaSourceUtils/>`
task invocations "loses" the item metadata, resulting in a warning:

    warning XA1020: At least one Java library is required for binding.
    Check that a Java library is included in the project and has the appropriate build action:
    'LibraryProjectZip' (for AAR or ZIP), 'EmbeddedJar', 'InputJar' (for JAR), or 'LibraryProjectProperties' (project.properties).

However, this *should* be the case on `main` *now*, but XA1020
*isn't* produced!

Conjecture: the problem is that c30d2aae added files to
`@(_JavaSourceJavadocXml)`, and this change "breaks" things, possibly
because it "unaligns" the Inputs & Outputs of the
`_ExtractJavadocsFromJavaSourceJars` target, disabling target batching.
Possibly related: https://docs.microsoft.com/en-us/visualstudio/msbuild/incremental-builds?view=vs-2022

> A target element can have both an `Inputs` attribute, which
> indicates what items the target expects as input, and an `Outputs`
> attribute, which indicates what items it produces as output.
> MSBuild attempts to find a 1-to-1 mapping between the values of
> these attributes. If a 1-to-1 mapping exists, MSBuild compares the
> time stamp of every input item to the time stamp of its
> corresponding output item. Output files that have no 1-to-1 mapping
> are compared to all input files.

By adding an entry to `@(_JavaSourceJavadocXml)` but *not*
`@(JavaSourceJar)`, there is no longer a 1-to-1 mapping.

Perhaps this is what causes the breakage?

Update `_CompileBindingJava` target so that it *doesn't* add files
to `@(_JavaSourceJavadocXml)`, and instead update the
`GenerateBindings` target so that it uses
`$(_AndroidIntermediateBindingClassesDocs)`.

Perhaps this will fix the `DotNetBuild()` unit test?