Revert "Revert "Ignore duplicate inputs in <JavaSourceUtils/>""
This reverts commit e525b0042403584359b2c1ac0ed30b0fffee6912.

Lol?

While "duplicate item groups" *should* exist, the *original*
rationale for removing duplicates in e59b8b03
was because *`java-source-utils`* doesn't like duplicates, and
we're trying to *add* an invocation of `java-source-utils`.

This might not be the best way -- [another suggestion][0] was to use
the [`<RemoveDuplicates/>` task][1] -- but it will at least
let us know if there are any other issues waiting for fixes.

[0]: https://discord.com/channels/732297728826277939/732297837953679412/982037330938454068
[1]: https://docs.microsoft.com/en-us/visualstudio/msbuild/removeduplicates-task?view=vs-2022