[build] Copy .NET build output to ref/runtime dirs (#7010)
How does one test incremental changes to xamarin-android for .NET6+?
Previously, every change would require re-creating the .NET packs:

	# make changes to src/Xamarin.Android.Build.Tasks
	% dotnet build src/Xamarin.Android.Build.Tasks/Xamarin.Android.Build.Tasks.csproj

	# re-create .NET packs
	% make pack-dotnet

	# can now use `./dotnet-local.*` to test changes.

The problem with this is that `make pack-dotnet` can take multiple
minutes to complete.  This slows down local testing.

Update the build system to instead place build artifacts within the
`bin\$(Configuration)\lib\packs` directory, and update the
`dotnet-local` scripts to set:

  * The `DOTNETSDK_WORKLOAD_MANIFEST_ROOTS` environment variable to
    `bin\$(Configuration)\lib\sdk-manifests`, and the

  * The `DOTNETSDK_WORKLOAD_PACK_ROOTS` environment variable to
    `bin\$(Configuration)\lib`.

This simplifies the .NET 6 test workflow to:

	./dotnet-local build Xamarin.Android.sln
	# Use of `dotnet-local` now includes in-tree changes.

The `bin\$(Configuration)\lib\packs` directory contains the content
required for our Ref and Runtime packs, and follows the directory
structure that the `dotnet workload` resolution process expects.
This in-build-tree structure is simplified compare to the CI
structure; commit distance and semver metadata isn't present.

A new `ConfigureLocalWorkload` target has been added to run a couple
of extra steps that are required to set up the local workload.  This
includes generating workload manifest, target, and FrameworkList files,
installing mono/runtime dependencies, and creating the templates pack.
It is run as part of the `make all` and `make jenkins` Make targets.

The `ConfigureLocalWorkload` target should only need to run once per
"clean build".  It is fine to run this target more frequently than
that, but it is only required when building from scratch or after one
of the following changes:

  * The structure of the android workload changes.
  * Files are added to either the Ref or Runtime packs.
  * Changes are made to the templates pack.
  * The dotnet/runtime dependency is updated.

A new `BuildDotNet` target has been added, which will only build one
API level.  This should be closer to `make all`, *not* `make jenkins`.
The `BuildDotNet` target also calls `ConfigureLocalWorkload`.