[XABT] Support testing with dotnet packs override
Adds support to the desktop MSBuild tests to find and use build output
in the `bin/$(Configuration)/lib/packs` folder.  This is an extension
of the changes in commits b1232a2b and c21e78ca which allowed local
builds to skip .nupkg packing and installing.

Some additional test clean up and refactoring is included.