[XABT] Support testing with dotnet packs override (#7096)
Adds support to the desktop MSBuild tests to find and use build output
in the `bin/$(Configuration)/lib/packs` folder.  This is an extension
of the changes in commits b1232a2b and c21e78ca which allowed local
builds to skip .nupkg packing and installing.

The `Xamarin.Android.Build.Tests` and `MSBuildDeviceIntegration` test
suites will now attempt to run against local build outputs by default,
and fall back to `dotnet/packs` or a legacy system install location if
a local build does not exist.