[Xamarin.Android.Build.Tasks] Handle Empty R.txt file.
Fixes https://i.azdo.io/1439405

The ManagedResourceParser was not handling the case where we do not
get four items per line. As a result it would crash with the following
error.

```
System.IndexOutOfRangeException: Index was outside of bounds of the array.
as Xamarin.And.Tasks.ManagedResourceParser.ProcessRtxtFile (System.String file)
```

If we do not get the expected four items from the line in the file we
should ignore that line. A unit test has been added to check that the
task does not fail when this happens.