[Xamarin.Android.Build.Tasks] Handle Empty R.txt file. (#7133)
Fixes: https://dev.azure.com/DevDiv/DevDiv/_workitems/edit/1439405

`ManagedResourceParser` would assume that `R.txt` would contain four
entries space-delimited entries per line, e.g.

	int layout main 0x7f030000

If `ManagedResourceParser` encountered a line that didn't have four
space-delimited entries, then it would crash:

	System.IndexOutOfRangeException: Index was outside of bounds of the array.
	   as Xamarin.And.Tasks.ManagedResourceParser.ProcessRtxtFile (System.String file)

We do not yet know of an "end-to-end" scenario that would cause
`R.txt` to have a line without four entries.

We *do* know how to force the error in a unit-test context: provide
an `R.txt` file that contains an empty line

Update `ManagedResourceParser` to verify that the number of values
returned from `string.Split()` is four, and ignore the line when that
is not the case.  This avoids the `IndexOutOfRangeException`.

A unit test has been added to check that the task does not fail when
this happens.